Initial commit of go-scamalytics application
This commit includes the initial version of the go-scamalytics command-line application, which allows users to check the fraud score of a list of IP addresses using scamalytics.com. The application is designed for security professionals and developers who need to quickly check the fraud score of multiple IP addresses and generate a report in CSV format. The app is written in Go and can be compiled for multiple platforms, including Windows, Linux, and macOS. This commit also includes a makefile for easy compilation of the application, and a README.md file with instructions on how to use and install the application.